issue10: split ImageModal into useImageGallery hook + pure component
