+++ src/hooks/index.ts
@@ -1,7 +1,5 @@
 export * from './useDogDetails'
-export * from './useDogList'
-export * from './useDogVariants'
-export * from './useDogDetails'
 export * from './useDogDescription'
 export * from './useDogList'
 export * from './useDogVariants'
+export * from './useImageGallery'
