+++ src/modules/DogSearch/ImageModal/ImageModal.tsx
@@ -1,5 +1,6 @@
-import { FC, useEffect, useState, useCallback } from 'react'
+import { FC } from 'react'
 import { RoundButton } from '../../../components'
+import { useImageGallery } from '../../../hooks'
 
 interface ImageModalProps {
     imageUrl: string
@@ -12,54 +13,8 @@ export const ImageModal: FC<ImageModalProps> = ({
     imageList,
     onClose,
 }) => {
-    const [currentIndex, setCurrentIndex] = useState(0)
-
-    useEffect(() => {
-        if (imageList) {
-            const index = imageList.indexOf(imageUrl)
-            if (index !== -1) {
-                setCurrentIndex(index)
-            }
-        }
-    }, [imageUrl, imageList])
-
-    const handlePrevious = useCallback(() => {
-        if (imageList && imageList.length > 0) {
-            setCurrentIndex(prev =>
-                prev === 0 ? imageList.length - 1 : prev - 1
-            )
-        }
-    }, [imageList])
-
-    const handleNext = useCallback(() => {
-        if (imageList && imageList.length > 0) {
-            setCurrentIndex(prev => (prev + 1) % imageList.length)
-        }
-    }, [imageList])
-
-    useEffect(() => {
-        const handleKeyDown = (e: KeyboardEvent) => {
-            if (e.key === 'Escape') {
-                onClose()
-            } else if (imageList && imageList.length > 0) {
-                if (e.key === 'ArrowLeft') {
-                    handlePrevious()
-                } else if (e.key === 'ArrowRight') {
-                    handleNext()
-                }
-            }
-        }
-        document.addEventListener('keydown', handleKeyDown)
-        document.body.style.overflow = 'hidden'
-        return () => {
-            document.removeEventListener('keydown', handleKeyDown)
-            document.body.style.overflow = ''
-        }
-    }, [handlePrevious, handleNext, imageList, onClose])
-
-    const displayImage =
-        imageList && imageList.length > 0 ? imageList[currentIndex] : imageUrl
-    const showArrows = imageList && imageList.length > 1
+    const { displayImage, showArrows, handlePrevious, handleNext } =
+        useImageGallery({ imageUrl, imageList, onClose })
 
     return (
         <div
